% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readme2_helper_fxns.R
\name{cleanme}
\alias{cleanme}
\title{cleanme}
\usage{
cleanme(my_text, finalEncoding = "ASCII")
}
\arguments{
\item{my_text}{Vector of character strings containing the raw document texts.}

\item{finalEncoding}{A character string indicating the desired encoding for the text vector (default = "ASCII")}
}
\value{
A vector of character strings with the processed texts, each token is separated by a space.
}
\description{
Standard preprocessing code for ASCII texts. Removes HTML tags, URLs, linebreaks. Converts standard emoticons to tokens. Removes non-informative punctuation.
}
