% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readme2_helper_fxns.R
\name{download_wordvecs}
\alias{download_wordvecs}
\title{download_wordvecs}
\usage{

  download_wordvecs(url = "http://gking-projects.iq.harvard.edu/files/glove.6B.200d.zip",
  targetDir = NULL)
}
\arguments{
\item{url}{URL of the word vector file. If file is a .zip archive, contents will be unzipped into target directory.}

\item{targetDir}{Path to target directory. If NULL, uses the install directory for the 'readme' package as a default.}
}
\value{
Local path to downloaded file.
}
\description{
Downloads and stores word vector files in the readme local directory.
}
