% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readme2_readme_historical.R
\name{readme0}
\alias{readme0}
\title{readme0
Internal}
\usage{
readme0(
  dtm,
  labeledIndicator,
  categoryVec,
  features = 15,
  nboot = 10,
  probWt = 1
)
}
\description{
readme0
Internal
}
\keyword{internal}
